feat(#60): 스터디 신청 철회 API 구현
에러핸들링을 활용하여 가입된 스터디 or 가입 신청한 스터디인 경우 SnackBar를 통해 철회 -> 변경 필요

+++ app/src/main/java/com/ssafy/neegongnaegong/data/datasource/network/NetworkStudiesDataSource.kt
@@ -29,5 +29,5 @@ interface NetworkStudiesDataSource {
 
     suspend fun applyStudies(studyGroupId: Long): Flow<Unit>
 
-    suspend fun cancelApplicationsStudies(studyGroupId: Int): Flow<Unit>
+    suspend fun cancelApplicationsStudies(studyGroupId: Long): Flow<Unit>
 }
